#9 - 기존의 batch 프로젝트 생성 시 애플리케이션 실행 되지 않고 에러 발생되어 프로젝트 구조 수정
todo: 현재 애플리케이션 실행 시 에러는 발생되지 않으나 실제 예제에 정의 해둔 job 이 제대로 실행 되지 않는 문제는 해결되지 않아 문제 해결 필요
