+++ src/main/java/com/fastcampus/pass/config/JobConfig.java
@@ -1,0 +1,10 @@
+package com.fastcampus.pass.config;
+
+import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableBatchProcessing
+public class JobConfig {
+
+}
