+++ src/main/java/com/fastcampus/pass/config/PassJobConfig.java
@@ -1,0 +1,45 @@
+package com.fastcampus.pass.config;
+
+import org.springframework.batch.core.Job;
+import org.springframework.batch.core.Step;
+import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
+import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
+import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
+import org.springframework.batch.repeat.RepeatStatus;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableBatchProcessing
+public class PassJobConfig {
+
+    public static final String JOB_NAME = "passJob";
+    private final JobBuilderFactory jobBuilderFactory;
+    private final StepBuilderFactory stepBuilderFactory;
+
+    @Autowired
+    public PassJobConfig(JobBuilderFactory jobBuilderFactory,
+                         StepBuilderFactory stepBuilderFactory) {
+        this.jobBuilderFactory = jobBuilderFactory;
+        this.stepBuilderFactory = stepBuilderFactory;
+    }
+
+    @Bean
+    public Job passJob() {
+        return jobBuilderFactory.get(JOB_NAME)
+                .start(passStep())
+                .build();
+    }
+
+    @Bean
+    public Step passStep() {
+        return stepBuilderFactory.get("passStep")
+                .tasklet((contribution, chunkContext) -> {
+                    System.out.println("==================== execute passTasklet ====================");
+                    return RepeatStatus.FINISHED;
+                })
+                .build();
+    }
+
+}
